+++ command/group_finish.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"strconv"
-	"time"
 	"trace/tracing"
 
 	"github.com/mitchellh/cli"
 	"github.com/spf13/pflag"
-	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -107,48 +104,4 @@ func (c *GroupFinishCommand) RunContext(ctx context.Context, args []string) erro
 	finishSpan(span, c.now())
 
 	return nil
-}
-
-func createSpan(tp trace.Tracer, traceParent string, props map[string]string) (trace.Span, error) {
-	nano := props["start"]
-	i, err := strconv.ParseInt(nano, 10, 64)
-	if err != nil {
-		return nil, err
-	}
-	start := time.Unix(0, i)
-
-	// cli carrier traceParent
-	ctx := tracing.WithTraceParent(context.Background(), traceParent)
-	_, span := tp.Start(ctx, props["name"], trace.WithTimestamp(start))
-
-	return span, nil
-}
-
-func finishSpan(span trace.Span, finish int64) {
-	span.End(trace.WithTimestamp(time.Unix(0, finish)))
-}
-
-func applyProps(span trace.Span, props map[string]string) {
-
-	delete(props, "name")
-	delete(props, "start")
-
-	attrs := tracing.AttributesFromMap(props)
-
-	span.SetAttributes(attrs...)
-}
-
-func applyStatus(span trace.Span, flag *pflag.Flag) {
-
-	if flag != nil && flag.Changed {
-
-		value := flag.Value.String()
-		if value == "unset" {
-			value = ""
-		}
-
-		span.SetStatus(codes.Error, value)
-	} else {
-		span.SetStatus(codes.Ok, "")
-	}
 }